+++ src/commands/pokemon/release.js
@@ -1,11 +1,12 @@
 const { getTrainer } = require('../../services/trainer');
 const { listPokemons } = require('../../services/pokemon');
 const { MAX_RELEASE } = require('../../config/trainerConfig');
-const { idFrom } = require("../../database/mongoHandler");
+const { idFrom } = require("../../utils/utils");
 const { buildPokemonListEmbed } = require('../../embeds/pokemonEmbeds');
 const { buildYesNoActionRow } = require('../../components/yesNoActionRow');
 const { eventNames } = require('../../config/eventConfig');
 const { setState } = require('../../services/state');
+const { calculateWorth } = require('../../utils/pokemonUtils');
 
 const release = async (user, pokemonIds) => {
     if (pokemonIds.length > MAX_RELEASE) {
@@ -21,9 +22,10 @@ const release = async (user, pokemonIds) => {
     if (toRelease.err) {
         return { err: toRelease.err };
     } else if (toRelease.data.length !== pokemonIds.length) {
-        console.log(toRelease.data)
         return { err: `You don't have all the pokemon you want to release!` };
     }
+
+    const totalWorth = calculateWorth(toRelease.data, null);
 
     const embed = buildPokemonListEmbed(trainer.data, toRelease.data, 1);
 
@@ -32,29 +34,34 @@ const release = async (user, pokemonIds) => {
         stateId: stateId,
     }
     const actionRow = buildYesNoActionRow(releaseData, eventNames.POKEMON_RELEASE, true);
-    
-    return { embeds: [embed], components: [actionRow], err: null };
+
+    const send = {
+        content: `Do you really want to release the following Pokemon for ₽${totalWorth}?`,
+        embeds: [embed],
+        components: [actionRow],
+    }
+    return { send: send, err: null };
 }
 
 const releaseMessageCommand = async (message) => {
     const pokemonIds = message.content.split(' ').slice(1);
-    const { embeds, components, err } = await release(message.author, pokemonIds);
+    const { send, err } = await release(message.author, pokemonIds);
     if (err) {
         await message.channel.send(`${err}`);
         return { err: err };
     } else {
-        await message.channel.send({ content: "Are you sure you want to release the following Pokemon?", embeds: embeds, components: components })
+        await message.channel.send(send);
     }
 }
 
 const releaseSlashCommand = async (interaction) => {
     const pokemonIds = interaction.options.getString('pokemonids').split(' ');
-    const { embeds, components, err } = await release(interaction.user, pokemonIds);
+    const { send, err } = await release(interaction.user, pokemonIds);
     if (err) {
         await interaction.reply(`${err}`);
         return { err: err };
     } else {
-        await interaction.reply({ content: "Are you sure you want to release the following Pokemon?", embeds: embeds, components: components });
+        await interaction.reply(send);
     }
 }