+++ src/handlers/eventHandler.js
@@ -1,6 +1,6 @@
 const { eventNames, eventConfig } = require("../config/eventConfig.js");
 const { logger } = require("../log");
-const { addExp } = require("../services/trainer");
+const { addExpAndMoney: addExpAndMoney } = require("../services/trainer");
 const path = require("path");
 
 const eventHandlers = {};
@@ -26,14 +26,14 @@ const handleEvent = async (interaction) => {
         res = await eventHandlers[eventName](interaction, data);
         if (res && res.err) {
             logger.warn(`Error executing event ${eventName}`);
-            logger.warn(res.err);
             return;
         }
-
-        // add exp
-        const exp = eventConfig[eventName].exp;
-        if (exp && exp > 0) {
-            const { level, err } = await addExp(interaction.user, exp);
+        
+        // add exp & money if possible
+        const exp = eventConfig[eventName].exp || 0;
+        const money = eventConfig[eventName].money || 0;
+        if (exp > 0 || money > 0) {
+            const { level, err } = await addExpAndMoney(interaction.user, exp, money);
             if (err) {
                 return;
             } else if (level) {